bzero

+++ src/ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: ikawamuk <ikawamuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/25 10:08:19 by ikawamuk          #+#    #+#             */
-/*   Updated: 2025/04/25 10:23:24 by ikawamuk         ###   ########.fr       */
+/*   Updated: 2025/04/25 10:27:38 by ikawamuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 	unsigned char	*p;
 
 	p = (unsigned char *)s;
