+++ src/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: ikawamuk <ikawamuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/25 10:02:06 by ikawamuk          #+#    #+#             */
-/*   Updated: 2025/04/25 10:11:36 by ikawamuk         ###   ########.fr       */
+/*   Updated: 2025/04/25 10:27:52 by ikawamuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	unsigned int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len] != '\0')
